requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
numpy==1.24.0